third_party/WebKit/LayoutTests/fast/events/click-with-large-negative-text-indent.html

<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<button id="testButton" style="text-indent: -99999999px;" onclick="btnClicked()">Move</button>
<p id="output"></p>
<script>
const testCase = async_test('Button is clicked');

const btnClicked = testCase.step_func_done(() => {});

testCase.step(() => {
  assert_own_property(window, 'eventSender');
  var testEle = document.getElementById('testButton');
  eventSender.mouseMoveTo(testEle.offsetLeft + testEle.offsetWidth / 2, testEle.offsetTop + testEle.offsetHeight / 2);
  eventSender.mouseDown();
  eventSender.mouseUp();
});
</script>
</body>
</html>
